+++ bluepysnap/edges/edge_population.py
@@ -27,9 +27,9 @@ from more_itertools import first
 
 from bluepysnap import query, utils
 from bluepysnap.circuit_ids import CircuitEdgeIds
+from bluepysnap.circuit_ids_types import IDS_DTYPE, CircuitEdgeId
 from bluepysnap.exceptions import BluepySnapError
 from bluepysnap.sonata_constants import DYNAMICS_PREFIX, ConstContainer, Edge
-from bluepysnap.utils import IDS_DTYPE, CircuitEdgeId
 
 
 def _is_empty(xs):
