+++ bluepysnap/nodes/node_population.py
@@ -69,6 +69,7 @@ from more_itertools import first
 
 from bluepysnap import query, utils
 from bluepysnap.circuit_ids import CircuitNodeIds
+from bluepysnap.circuit_ids_types import CircuitNodeId
 from bluepysnap.exceptions import BluepySnapError
 from bluepysnap.node_sets import NodeSet
 from bluepysnap.sonata_constants import DYNAMICS_PREFIX, ConstContainer, Node
@@ -429,7 +430,7 @@ class NodePopulation:
         else:
             result = utils.ensure_list(group)
             # test if first value is a CircuitNodeId all values are all CircuitNodeId
-            if isinstance(first(result, None), utils.CircuitNodeId):
+            if isinstance(first(result, None), CircuitNodeId):
                 try:
                     result = [cid.id for cid in result if cid.population == self.name]
                 except AttributeError as e:
@@ -526,7 +527,7 @@ class NodePopulation:
         if group is not None:
             if isinstance(group, (int, np.integer)):
                 self._check_id(group)
-            elif isinstance(group, utils.CircuitNodeId):
+            elif isinstance(group, CircuitNodeId):
                 group = self.ids(group, raise_missing_property=raise_missing_property)[0]
             else:
                 group = self.ids(group, raise_missing_property=raise_missing_property)
