+++ tests/test_edges.py
@@ -12,8 +12,8 @@ import bluepysnap.edges as test_module
 from bluepysnap.bbp import Synapse
 from bluepysnap.circuit import Circuit
 from bluepysnap.circuit_ids import CircuitEdgeIds, CircuitNodeIds
+from bluepysnap.circuit_ids_types import IDS_DTYPE, CircuitEdgeId, CircuitNodeId
 from bluepysnap.exceptions import BluepySnapError
-from bluepysnap.utils import IDS_DTYPE, CircuitEdgeId, CircuitNodeId
 
 from utils import PICKLED_SIZE_ADJUSTMENT, TEST_DATA_DIR
