+++ tests/test_frame_report.py
@@ -9,9 +9,9 @@ import pytest
 import bluepysnap.frame_report as test_module
 from bluepysnap.bbp import Cell
 from bluepysnap.circuit_ids import CircuitNodeIds
+from bluepysnap.circuit_ids_types import IDS_DTYPE, CircuitNodeId
 from bluepysnap.exceptions import BluepySnapError
 from bluepysnap.simulation import Simulation
-from bluepysnap.utils import IDS_DTYPE, CircuitNodeId
 
 from utils import TEST_DATA_DIR
